BE: make sure only im able to create, update, destroy, items

+++ src/server/deleteProduct.ts
@@ -1,10 +1,17 @@
 import { server$ } from "@builder.io/qwik-city";
 import { getDb } from "~/db/mongodb";
 import Types from "mongodb";
+import { Session } from "@auth/qwik";
 
 export const deleteProduct = server$(async function (id: string )  {
   try {
     const uri = this.env.get("MONGO_URI") || "";
+    const session: Session | null = this.sharedMap.get("session");
+
+    if (!session || session.user?.email !== "ts22082@gmail.com") {
+      return { success: false };
+    }
+
     const db = await getDb(uri);
     const filter = { _id: new Types.ObjectId(id) };
     const result = await db.collection("products").deleteOne(filter);
